refactor(lexer): split `find` module into multiple files (#26506)
Pure refactor.

Split `find` module into parts.

- `find` directory contains find primitives (`find1`, `find2` etc) and specialized `find_*` functions.
- `scan` directory contains `scan_*` functions.
- `chunk` module contains primitives for reading and manipulating chunks (`load256`, `mm`, `veq` etc).

In `find` and `scan`, AVX2 and generic code is split into separate files, which both export the same functions.

Each component is now small enough to read as a whole.

This is pure refactor in the strictest sense. The ASM the crate compiles to remains identical to before this PR.
