+++ crates/oxc_lexer/src/pipeline/carve/common.rs
@@ -9,8 +9,8 @@ use crate::{
 use super::super::{
     BCOM, LCOM, REGEX, STR,
     bitmap::{bm_clear, bm_clear_range, bm_get, bm_next0, bm_set},
-    find::{scan_block_comment, scan_line_comment, scan_quoted, scan_regex, scan_tmpl_text},
     regex_div::prev_is_regex,
+    scan::{scan_block_comment, scan_line_comment, scan_quoted, scan_regex, scan_tmpl_text},
 };
 
 /// Lex the string literal opening at `s`. Returns the resume index. Shared
