+++ crates/oxc_lexer/src/pipeline/carve/jsx/angle_brackets.rs
@@ -6,8 +6,9 @@ use crate::{
 
 use super::super::super::{
     bitmap::bm_get,
-    find::{find_line_terminator, scan_block_comment, unicode_ws_len},
+    find::{find_line_terminator, unicode_ws_len},
     regex_div::{jsx_site_is_expression, ts_type_region_open, type_parameter_list_head},
+    scan::scan_block_comment,
 };
 
 const FN_TYPE_SCAN_CAP: usize = 1 << 16;
