+++ crates/oxc_lexer/src/pipeline/carve/jsx/mod.rs
@@ -10,9 +10,10 @@ use super::super::{
     bitmap::{bm_clear, bm_clear_range, bm_set},
     find::{
         find_jsx_tag, find_jsx_text, find_line_terminator, find_opener, find_opener_jsx5,
-        find_opener_jsx7, find_opener6, find1, find2, scan_block_comment,
+        find_opener_jsx7, find_opener6, find1, find2,
     },
     regex_div::{bm_prev_sig, prev_is_regex},
+    scan::scan_block_comment,
 };
 
 use super::common::{
