+++ crates/oxc_lexer/src/pipeline/coalesce/mod.rs
@@ -8,8 +8,8 @@ use crate::{
 use super::{
     NUM,
     bitmap::{bm_clear, bm_clear_range, bm_get, bm_next0, bm_set},
-    find::scan_number,
     regex_div::{gt_run_split, lt_run_split},
+    scan::scan_number,
 };
 
 mod keywords;
