+++ crates/oxc_lexer/src/pipeline/find/mod.rs
@@ -1,0 +1,19 @@
+#[cfg(all(target_arch = "x86_64", target_feature = "avx2", target_feature = "bmi2"))]
+mod avx2;
+#[cfg(all(target_arch = "x86_64", target_feature = "avx2", target_feature = "bmi2"))]
+use avx2::finder;
+#[cfg(all(target_arch = "x86_64", target_feature = "avx2", target_feature = "bmi2"))]
+pub(super) use avx2::{find1, find2, find3, find4};
+
+#[cfg(not(all(target_arch = "x86_64", target_feature = "avx2", target_feature = "bmi2")))]
+mod generic;
+#[cfg(not(all(target_arch = "x86_64", target_feature = "avx2", target_feature = "bmi2")))]
+use generic::finder;
+#[cfg(not(all(target_arch = "x86_64", target_feature = "avx2", target_feature = "bmi2")))]
+pub(super) use generic::{find1, find2, find3, find4};
+
+mod common;
+pub(super) use common::{
+    find_jsx_tag, find_jsx_text, find_line_terminator, find_opener, find_opener_jsx5,
+    find_opener_jsx7, find_opener6, find_regex, find_tmpl, unicode_ws_len,
+};
